Drop the port-forwarding dream.
On a happier note, got the rbac situation settled so the proxy has a service account which allows it to read the endpoints.  So the proxy works, but before the k8s file you must add a ClusterRole for yourself so you have the cluster-admin role, then you can set the ServiceAccount on the deployment.